fix api breaking change
